ref: Change implementation of StockManagementWindow - extract views into two separate UserControls. Create new DeliveryService to menage between stock and delivery. Remove unnecessary files
