+++ POS/ViewModels/WarehouseFunctions/WarehouseFunctionsViewModel.cs
@@ -24,8 +24,9 @@ namespace POS.ViewModels.WarehouseFunctions
         }
 
         public ICommand LoadRunningOutOfIngredientsCommand { get; }
+        public ICommand OpenProductManagementWindowCommand { get; }
         public ICommand OpenStockManagementWindowCommand { get; }
-        public ICommand OpenCreateDeliveryWindowCommand { get; }
+        //public ICommand OpenCreateDeliveryWindowCommand { get; }
 
         public WarehouseFunctionsViewModel(IngredientService ingredientService, NavigationService navigationService)
         {
@@ -33,8 +34,9 @@ namespace POS.ViewModels.WarehouseFunctions
             _navigationService = navigationService;
 
             LoadRunningOutOfIngredientsCommand = new RelayCommandAsync(LoadRunningOutOfIngredientsAsync);
-            OpenStockManagementWindowCommand = new RelayCommand<StockManagementWindow>(OpenWindow);
-            OpenCreateDeliveryWindowCommand = new RelayCommand<CreateDeliveryWindow>(OpenWindow);
+            OpenProductManagementWindowCommand = new RelayCommand<ProductManagementWindow>(OpenWindow);
+            OpenStockManagementWindowCommand = new RelayCommand<StockAndDeliveryManagementWindow>(OpenWindow);
+            //OpenCreateDeliveryWindowCommand = new RelayCommand<CreateDeliveryWindow>(OpenWindow);
         }
 
         private async Task LoadRunningOutOfIngredientsAsync()
