+++ POS/Views/UserControls/WarehouseFunctions/CreateDeliveryUserControl.xaml
@@ -1,0 +1,95 @@
+﻿<UserControl x:Class="POS.Views.UserControls.WarehouseFunctions.CreateDeliveryUserControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:fa="http://schemas.fontawesome.io/icons/"
+             xmlns:warehouseFunctions="clr-namespace:POS.ViewModels.WarehouseFunctions"
+             d:DataContext="{d:DesignInstance Type=warehouseFunctions:CreateDeliveryViewModel}"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+
+    <Grid>
+        <Border BorderThickness="0,0,1,0" BorderBrush="Black" d:IsLocked="True"/>
+        <Grid Grid.Column="0" Margin="25">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="50"/>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="70"/>
+            </Grid.RowDefinitions>
+
+            <Border Grid.Row="0" BorderThickness="1,1,1,0" BorderBrush="Black" Background="#8ea7e9" CornerRadius="8,8,0,0" d:IsLocked="True">
+                <Grid>
+                    <TextBlock Text="LISTA PRODUKTÓW DO ZAMÓWIENIA" VerticalAlignment="Center" HorizontalAlignment="Center" Foreground="White" FontSize="16"/>
+                </Grid>
+            </Border>
+
+            <Border Grid.Row="1" BorderThickness="1" BorderBrush="Black" CornerRadius="0,0,0,0" d:IsLocked="True">
+                <Grid>
+                    <DataGrid ItemsSource="{Binding DeliveryObservableCollection}"
+                      Style="{DynamicResource DataGridStyle3}"
+                      RowStyle="{DynamicResource DataGridRowStyle3}"
+                      CellStyle="{DynamicResource DataGridCellStyle3}"
+                      ColumnHeaderStyle="{DynamicResource DataGridColumnHeaderStyle3}"
+                      BorderThickness="0"
+                      IsReadOnly="True">
+                        <DataGrid.Columns>
+                            <DataGridTextColumn Header="Składnik" Width="3*" Binding="{Binding Ingredient.Name}"/>
+                            <DataGridTextColumn Header="Opakowanie" Width="2*" Binding="{Binding Ingredient.Package}"/>
+                            <DataGridTextColumn Header="Jednostka" Width="2*" Binding="{Binding Ingredient.Unit}"/>
+
+                            <DataGridTemplateColumn Header="" Width="0.5*">
+                                <DataGridTemplateColumn.CellTemplate>
+                                    <DataTemplate>
+                                        <Button Style="{StaticResource AddDeleteItemButton}"
+                                        HorizontalAlignment="Right" 
+                                        Command="{Binding DataContext.IncreaseIngredientQuantityCommand, RelativeSource={RelativeSource AncestorType=UserControl}}"
+                                        CommandParameter="{Binding }">
+                                            <fa:ImageAwesome Style="{StaticResource AddItemIcon}" Icon="Plus" />
+                                        </Button>
+                                    </DataTemplate>
+                                </DataGridTemplateColumn.CellTemplate>
+                            </DataGridTemplateColumn>
+
+                            <DataGridTextColumn Header="Ilość" Width="0.5*" Binding="{Binding Quantity}"/>
+
+                            <DataGridTemplateColumn Header="" Width="0.5*">
+                                <DataGridTemplateColumn.CellTemplate>
+                                    <DataTemplate>
+                                        <Button Style="{StaticResource AddDeleteItemButton}"
+                                        HorizontalAlignment="Right" 
+                                        Command="{Binding DataContext.DeleteIngredientFromDeliveryCommand, RelativeSource={RelativeSource AncestorType=UserControl}}"
+                                        CommandParameter="{Binding Ingredient}">
+                                            <fa:ImageAwesome Style="{StaticResource DeleteItemIcon}" Icon="Minus" />
+                                        </Button>
+                                    </DataTemplate>
+                                </DataGridTemplateColumn.CellTemplate>
+                            </DataGridTemplateColumn>
+                        </DataGrid.Columns>
+                    </DataGrid>
+                </Grid>
+            </Border>
+
+            <Border Grid.Row="2" BorderThickness="1,0,1,1" BorderBrush="Black" Background="#8ea7e9" CornerRadius="0,0,8,8" d:IsLocked="True">
+                <Grid>
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition Width="3*"/>
+                        <ColumnDefinition Width="3*"/>
+                    </Grid.ColumnDefinitions>
+
+                    <Button Command="{Binding GenerateDeliveryCommand}" Grid.Column="0" Style="{StaticResource BasicButtonGreen}">
+                        <StackPanel Orientation="Horizontal">
+                            <TextBlock Text="GENERUJ ZAMÓWIENIE" VerticalAlignment="Center" HorizontalAlignment="Center" Foreground="White" FontSize="16"/>
+                        </StackPanel>
+                    </Button>
+
+                    <Button Command="{Binding CancelDeliveryCommand}"  Grid.Column="1" Style="{StaticResource BasicButtonRed}">
+                        <StackPanel Orientation="Horizontal">
+                            <TextBlock Text="ANULUJ ZAMÓWIENIE" VerticalAlignment="Center" HorizontalAlignment="Center" Foreground="White" FontSize="16"/>
+                        </StackPanel>
+                    </Button>
+                </Grid>
+            </Border>
+        </Grid>
+    </Grid>
+</UserControl>
