+++ POS/Views/Windows/WarehouseFunctions/ProductManagementWindow.xaml
@@ -1,15 +1,15 @@
-﻿<Window x:Class="POS.Views.Windows.WarehouseFunctions.StockManagementWindow"
+﻿<Window x:Class="POS.Views.Windows.WarehouseFunctions.ProductManagementWindow"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
       xmlns:userControls="clr-namespace:POS.Views.UserControls.WarehouseFunctions"
       xmlns:warehouseFunctions="clr-namespace:POS.ViewModels.WarehouseFunctions"
-      d:DataContext="{d:DesignInstance Type=warehouseFunctions:StockManagementViewModel}"
+      d:DataContext="{d:DesignInstance Type=warehouseFunctions:ProductManagementViewModel}"
       mc:Ignorable="d" 
       xmlns:fa="http://schemas.fontawesome.io/icons/"
       d:DesignHeight="768" d:DesignWidth="1366"
-      Title="StockManagement" Background="#FFD9D9D9" WindowState="Maximized" WindowStyle="None" FontFamily="Lato" FontSize="16">
+      Title="ProductManagementWindow" Background="#FFD9D9D9" WindowState="Maximized" WindowStyle="None" FontFamily="Lato" FontSize="16">
 
     <Grid>
         <Grid.ColumnDefinitions>
@@ -27,9 +27,9 @@
         <!--#endregion-->
 
         <GridSplitter Grid.Column="0" Width="1" Foreground="Black" Background="Black"/>
-        
+
         <!--#region middle-->
-        
+
         <Grid Grid.Column="1">
             <userControls:AddEditDeleteProductUserControl/>
         </Grid>
@@ -37,7 +37,7 @@
         <!--#endregion-->
 
         <GridSplitter Grid.Column="1" Width="1" Foreground="Black" Background="Black"/>
-        
+
         <!--#region right-->
         <Grid Grid.Column="2">
             <Grid.RowDefinitions>
@@ -65,8 +65,8 @@
                 </Grid.ColumnDefinitions>
 
                 <Button Command="{Binding OpenMainWindowCommand}"
-                        Style="{StaticResource HomeButton}" 
-                        Grid.Column="1" Margin="25">
+                    Style="{StaticResource HomeButton}" 
+                    Grid.Column="1" Margin="25">
                     <fa:ImageAwesome Icon="Home" Height="24" Foreground="White"/>
                 </Button>
             </Grid>
