+++ POS/Views/Windows/WarehouseFunctions/StockAndDeliveryManagmentWindow.xaml
@@ -1,0 +1,78 @@
+﻿<Window x:Class="POS.Views.Windows.WarehouseFunctions.StockAndDeliveryManagementWindow"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:userControls="clr-namespace:POS.Views.UserControls.WarehouseFunctions"
+      xmlns:warehouseFunctions="clr-namespace:POS.ViewModels.WarehouseFunctions"
+      xmlns:helpers="clr-namespace:POS.Helpers"
+      xmlns:fa="http://schemas.fontawesome.io/icons/"
+      d:DataContext="{d:DesignInstance Type=warehouseFunctions:StockAndDeliveryManagementViewModel}"
+      mc:Ignorable="d" 
+      d:DesignHeight="768" d:DesignWidth="1366"
+      Title="StockManagement" Background="#FFD9D9D9" WindowState="Maximized" WindowStyle="None" FontFamily="Lato" FontSize="16">
+
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*"/>
+            <ColumnDefinition Width="*"/>
+        </Grid.ColumnDefinitions>
+
+        <!--#region left-->
+
+        <Grid Grid.Column="0">
+            <userControls:CreateDeliveryUserControl/>
+        </Grid>
+
+        <!--#endregion-->
+
+        <GridSplitter Grid.Column="0" Width="1" Foreground="Black" Background="Black"/>
+
+        <!--#region right-->
+
+        <Grid Grid.Column="1">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="1*"/>
+                <RowDefinition Height="10*"/>
+                <RowDefinition Height="1*"/>
+            </Grid.RowDefinitions>
+
+            <Border Grid.Row="0" BorderThickness="0,0,0,1" BorderBrush="Black" Background="#8ea7e9" d:IsLocked="True">
+                <Grid Background="#8ea7e9">
+                    <TextBox Text="{Binding SearchPhrase, UpdateSourceTrigger=PropertyChanged, Delay=100}"
+                             Style="{StaticResource SearchBar}"
+                             Margin="20,0,0,0" 
+                             Width="360" 
+                             Height="42"
+                             helpers:TextBoxHelper.Placeholder="{Binding Placeholder}">
+                        <TextBox.Resources>
+                            <Style TargetType="{x:Type Border}">
+                                <Setter Property="CornerRadius" Value="16"/>
+                            </Style>
+                        </TextBox.Resources>
+                    </TextBox>
+
+                    <Border BorderBrush="#7286d3" BorderThickness="1" Width="200" Height="42" HorizontalAlignment="Right" Margin="0,0,25,0" Background="#FF7286D3" CornerRadius="8">
+                        <StackPanel VerticalAlignment="Center" HorizontalAlignment="Center">
+                            <TextBlock Text="{Binding LoggedInUserName, FallbackValue='Imię Nazwisko'}" Foreground="White"/>
+                        </StackPanel>
+                    </Border>
+                </Grid>
+            </Border>
+
+            <userControls:StockManagementUserControl Grid.Row="1" Margin="25"/>
+
+            <Grid Grid.Row="2">
+                <Button Command="{Binding OpenMainWindowCommand}"
+                        Style="{StaticResource HomeButton}"
+                        HorizontalAlignment="Right"
+                        Margin="0, 0, 25, 25"
+                        MinWidth="100">
+                    <fa:ImageAwesome Icon="Home" Height="24" Foreground="White"/>
+                </Button>
+            </Grid>
+        </Grid>
+
+        <!--#endregion-->
+    </Grid>
+</Window>
